+++ app/Http/Controllers/Dono/CupomController.php
@@ -2,7 +2,6 @@
 
 namespace App\Http\Controllers\Dono;
 
-use App\Http\Controllers\Concerns\AuthorizesSalao;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\StoreCupomRequest;
 use App\Http\Requests\UpdateCupomRequest;
@@ -10,27 +9,32 @@ use App\Models\Cupom;
 
 class CupomController extends Controller
 {
-    use AuthorizesSalao;
-
     public function index()
     {
+        $this->authorize('viewAny', Cupom::class);
+
         $salao = auth()->user()->salao;
         $cupons = $salao->cupons()->orderByDesc('created_at')->paginate(20);
+
         return view('dono.cupons.index', compact('cupons'));
     }
 
     public function create()
     {
+        $this->authorize('create', Cupom::class);
+
         return view('dono.cupons.create');
     }
 
     public function store(StoreCupomRequest $request)
     {
+        $this->authorize('create', Cupom::class);
+
         $data = $request->validated();
-        $data['codigo']    = strtoupper($data['codigo']);
-        $data['salao_id']  = auth()->user()->salao_id;
+        $data['codigo'] = strtoupper($data['codigo']);
+        $data['salao_id'] = auth()->user()->salao_id;
         $data['uso_atual'] = 0;
-        $data['ativo']     = $request->boolean('ativo', true);
+        $data['ativo'] = $request->boolean('ativo', true);
 
         Cupom::create($data);
 
@@ -41,17 +45,18 @@ class CupomController extends Controller
 
     public function edit(Cupom $cupom)
     {
-        $this->authorizeSalaoOwnership($cupom);
+        $this->authorize('update', $cupom);
+
         return view('dono.cupons.edit', compact('cupom'));
     }
 
     public function update(UpdateCupomRequest $request, Cupom $cupom)
     {
-        $this->authorizeSalaoOwnership($cupom);
+        $this->authorize('update', $cupom);
 
         $data = $request->validated();
         $data['codigo'] = strtoupper($data['codigo']);
-        $data['ativo']  = $request->boolean('ativo');
+        $data['ativo'] = $request->boolean('ativo');
 
         $cupom->update($data);
 
@@ -62,8 +67,9 @@ class CupomController extends Controller
 
     public function destroy(Cupom $cupom)
     {
-        $this->authorizeSalaoOwnership($cupom);
+        $this->authorize('delete', $cupom);
         $cupom->delete();
+
         return back()->with('success', 'Cupom excluído.');
     }
 }
